# docs/openapi.yaml
openapi: 3.1.0
info:
    title: Rhapsode
    version: 0.1.9
paths:
...
    /engines/{engine}:
        delete:
            operationId: uninstallEngine
            description: Only an engine this API installed. One the operator configured is theirs to remove.
            parameters:
                - name: engine
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "204":
                    description: No content
                "403":
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "404":
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "409":
                    description: Conflict
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
...
    /engines/{engine}/install:
        post:
            operationId: installEngine
            parameters:
                - name: engine
                  in: path
                  required: true
                  schema:
                      type: string
                - name: pull
                  in: query
                  required: false
                  schema:
                      type: string
                - name: accept
                  in: query
                  required: false
                  schema:
                      type: string
            responses:
                "202":
                    description: Response 202
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/InstallJob"
                "400":
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "403":
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "404":
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "409":
                    description: Conflict
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
    /engines/{engine}/reinstall:
        post:
            operationId: reinstallEngine
            parameters:
                - name: engine
                  in: path
                  required: true
                  schema:
                      type: string
                - name: accept
                  in: query
                  required: false
                  schema:
                      type: string
            responses:
                "202":
                    description: Response 202
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/InstallJob"
                "400":
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "403":
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "404":
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "409":
                    description: Conflict
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
    /engines/{engine}/pull:
        post:
            operationId: pullEngine
            parameters:
                - name: engine
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PullRequest"
            responses:
                "202":
                    description: Response 202
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/InstallJob"
                "403":
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "404":
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "409":
                    description: Conflict
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
    /installs:
        get:
            operationId: installJobs
            responses:
                "200":
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/InstallJob"
                "403":
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
    /installs/outdated:
        post:
            operationId: reinstallOutdated
            responses:
                "202":
                    description: Response 202
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ReinstallOutdated"
                "403":
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
    /installs/{job}:
        get:
            operationId: installJob
            parameters:
                - name: job
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/InstallJob"
                "403":
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
                "404":
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBody"
...
components:
    schemas:
...
        ErrorDetail:
            type: object
            properties:
                code:
                    type: string
                    enum:
                        - bad_request
                        - unknown_engine
                        - unknown_voice
                        - unsupported
                        - model_unavailable
                        - oom
                        - overloaded
                        - internal
                        - forbidden
                        - conflict
                    description: |-
                        A closed set that § 9 grows at the end. A reader that meets a code from a newer contract must
                        not lose the envelope over it: `message` and `retryable` are the two fields that decide what
                        the caller does next, and they parse fine. The core falls back to `internal` for the code and
                        keeps the rest, rather than reporting a parse failure in place of the real error.
                message:
                    type: string
                retryable:
                    type: boolean
            required:
                - code
                - message
                - retryable
            description: |-
                `retryable` is a field rather than something the client infers from the status, because the
                distinction that matters is between "this request was wrong" and "this request was fine and the
                server was not". A caller that conflates them either retries a permanent failure forever or
                discards work that would have succeeded on the next pass.
        ErrorBody:
            type: object
            properties:
                error:
                    $ref: "#/components/schemas/ErrorDetail"
            required:
                - error
...
        InstallJob:
            type: object
            properties:
                id:
                    type: string
                engine:
                    type: string
                kind:
                    type: string
                    enum:
                        - install
                        - pull
                        - reinstall
                variant:
                    type: string
                    description: What a pull fetches, or an install fetches in step 5.
                state:
                    type: string
                    enum:
                        - queued
                        - running
                        - succeeded
                        - failed
                step:
                    type: string
                    enum:
                        - venv
                        - packages
                        - verify
                        - register
                        - weights
                createdAt:
                    type: string
                    description: ISO 8601, UTC.
                startedAt:
                    type: string
                finishedAt:
                    type: string
                error:
                    $ref: "#/components/schemas/ErrorDetail"
                    description: Present exactly when `state` is `failed`.
            required:
                - id
                - engine
                - kind
                - state
                - createdAt
        ReinstallSkipped:
            type: object
            properties:
                engine:
                    type: string
                reason:
                    type: string
                    enum:
                        - licence
                        - busy
                        - uncatalogued
                    description: Needs `accept`; has a job already; the catalog no longer has it.
            required:
                - engine
                - reason
            description: |-
                An outdated engine `POST /installs/outdated` did not queue, and why: each is one a single
                reinstall would have to ask somebody about. § 10.
        ReinstallOutdated:
            type: object
            properties:
                jobs:
                    type: array
                    items:
                        $ref: "#/components/schemas/InstallJob"
                    description: One reinstall per outdated engine this API installed, in id order.
                skipped:
                    type: array
                    items:
                        $ref: "#/components/schemas/ReinstallSkipped"
            required:
                - jobs
                - skipped
        PullRequest:
            type: object
            properties:
                variant:
                    type: string
                    description: Absent means the engine's default variant.
...